Decided there was no need for useContext hook
