+++ cad/client/src/components/relatedComparison/OutfitSection.jsx
@@ -1,101 +1,59 @@
-import React, { useState, useEffect } from 'react';
-// import OutfitCard from './OutfitCard';
-// import HorizontalScroller from './HorizontalScroller';
-import { OutfitProvider } from '../../OutfitContext';
-import OutfitItems from './OutfitItems';
-
-const DEFAULT_PRODUCTS = [{
-  id: 1,
-  name: 'Camo Onesie',
-  category: 'Jackets',
-  defaultPrice: '140',
-  rating: 3,
-},
-{
-  id: 2,
-  name: 'Camo Onesie',
-  category: 'Jackets',
-  defaultPrice: '140',
-  rating: 3,
-},
-{
-  id: 3,
-  name: 'Camo Onesie',
-  category: 'Jackets',
-  defaultPrice: '140',
-  rating: 3,
-},
-{
-  id: 4,
-  name: 'Camo Onesie',
-  category: 'Jackets',
-  defaultPrice: '140',
-  rating: 3,
-},
-{
-  id: 5,
-  name: 'Camo Onesie',
-  category: 'Jackets',
-  defaultPrice: '140',
-  rating: 3,
-},
-{
-  id: 6,
-  name: 'Camo Onesie',
-  category: 'Jackets',
-  defaultPrice: '140',
-  rating: 3,
-},
-{
-  id: 7,
-  name: 'Camo Onesie',
-  category: 'Jackets',
-  defaultPrice: '140',
-  rating: 3,
-}, {
-  id: 7,
-  name: 'Camo Onesie',
-  category: 'Jackets',
-  defaultPrice: '140',
-  rating: 3,
-}, {
-  id: 7,
-  name: 'Camo Onesie',
-  category: 'Jackets',
-  defaultPrice: '140',
-  rating: 3,
-}, {
-  id: 7,
-  name: 'Camo Onesie',
-  category: 'Jackets',
-  defaultPrice: '140',
-  rating: 3,
-}, {
-  id: 7,
-  name: 'Camo Onesie',
-  category: 'Jackets',
-  defaultPrice: '140',
-  rating: 3,
-}];
+import React, { useState, useEffect, useContext } from 'react';
+import ProductContext from '../../ProductContext';
+import OutfitCard from './OutfitCard';
+import HorizontalScroller from './HorizontalScroller';
 
 const OutfitSection = () => {
-  // const [items, setItems] = useState(DEFAULT_PRODUCTS);
+  const [outfitItems, setOutfitItems] = useState([]);
+  const { productData } = useContext(ProductContext);
 
-  // const handleAddClick = () => {
-  //   // update items in outfit here w/ cur item
-  // };
+  useEffect(() => {
+    const outfitItemsData = JSON.parse(localStorage.getItem('outfitItems'));
+    if (outfitItemsData) {
+      setOutfitItems(outfitItemsData);
+    }
+  }, []);
 
-  // const handleRemoveClick = (itemId) => {
-  //   console.log(itemId);
-  //   setItems(items.filter((item) => item.id !== itemId));
-  // };
+  useEffect(() => {
+    localStorage.setItem('outfitItems', JSON.stringify(outfitItems));
+  }, [outfitItems]);
+
+  const addToOutfit = (newItem) => {
+    const hasItem = outfitItems.some((item) => item.id === newItem.id);
+    if (!hasItem) {
+      setOutfitItems((prevItems) => [...prevItems, newItem]);
+    }
+  };
+
+  const removeFromOutfit = (id) => {
+    setOutfitItems((prevItems) => prevItems.filter((item) => item.id !== id));
+  };
 
   return (
     <div className='outfit-container'>
       <h3>YOUR OUTFIT</h3>
-      <OutfitProvider>
-        <OutfitItems />
-      </OutfitProvider>
+      <HorizontalScroller>
+        <button className='add-outfit-btn' type='button' label='add-outfit-item' onClick={() => addToOutfit(productData)}>
+          <div className='product-card-container'>
+            +
+          </div>
+        </button>
+        {outfitItems.map(({
+        // eslint-disable-next-line camelcase
+          id, name, category, default_price, rating,
+        }) => (
+          <OutfitCard
+            handleRemoveClick={removeFromOutfit}
+            key={id}
+            id={id}
+            name={name}
+            category={category}
+          // eslint-disable-next-line camelcase
+            defaultPrice={default_price}
+            rating={rating}
+          />
+        ))}
+      </HorizontalScroller>
     </div>
   );
 };
